+++ src/wallet/wallet.module.ts
@@ -6,11 +6,19 @@ import { WalletPgRepository } from './repositories/wallet-pg.repository';
 import { WalletController } from './controller/wallet.controller';
 import { EncryptModule } from '../encrypt/encrypt.module';
 import { GetWalletService } from './services/get-wallet.service';
+import { WalletCommand } from '@app/wallet/cli/wallet.command';
+import { ExportWalletCommand } from '@app/wallet/cli/export-wallet.command';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Wallet]), EncryptModule],
   controllers: [WalletController],
-  providers: [CreateWalletService, WalletPgRepository, GetWalletService],
+  providers: [
+    CreateWalletService,
+    WalletPgRepository,
+    GetWalletService,
+    WalletCommand,
+    ExportWalletCommand,
+  ],
   exports: [GetWalletService, CreateWalletService],
 })
 export class WalletModule {}